feat : update logout
