--- src/shared/UserProvider.jsx
@@ -1,30 +1,0 @@
-import React from 'react';
-import { useState, useEffect } from 'react';
-import supabase from './supabaseClient';
-import { createContext } from 'react';
-
-//로그인, 로그아웃 추적
-
-export const UserContext = createContext();
-
-const UserProvider = ({ children }) => {
-  const [user, setUser] = useState(null);
-
-  useEffect(() => {
-    const {
-      data: { subscription }
-    } = supabase.auth.onAuthStateChange((session) => {
-      if (session) {
-        setUser(session.user);
-      } else {
-        setUser(null);
-      }
-    });
-
-    return () => subscription.unsubscribe();
-  }, []);
-
-  return <UserContext.Provider value={{ user, setUser }}>{children}</UserContext.Provider>;
-};
-
-export default UserProvider;
